+++ metadata.rb
@@ -1,8 +1,8 @@
-name             'opsworks-solr-read'
+name             'opsworks-solr'
 maintainer       'Sunil Sayala'
-maintainer_email 'sunil@'
+maintainer_email 'sunil.s@jobaline.com'
 license          'All rights reserved'
-description      'Installs/Configures RabbitMQ on Amazon Opsworks'
+description      'Installs/Configures Solr on Amazon Opsworks'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'