+++ recipes/logrotate.rb
@@ -1,0 +1,6 @@
+Chef::Log.info("updating logrotate configuration...")
+
+cookbook_file "rabbitmq-server" do
+  path "/etc/logrotate.d/rabbitmq-server"
+  action :create
+end